Refactored WebSocketServiceHost.cs, separated the WebSocketServiceHost<TBehavior> class into a new file
